[Scripts] Add a way to find what packages a script requires
Currently only works on systems with pacman as the package manager. It
will be simple to extract the distro-specific functionality.

This also requires that you actually already have all those packages
installed to begin with. So it would be used to keep a cache of needed
packages in git and when the dotfiles are pulled to a new system you
can run a command that will install all the packages in the cache.