[kaltemplate] Fix erroneous newline when using the dot_user macro
`whoami` outputs a newline and there doesn't seem to be an option to
change that. We can just pipe it into something like `tr -d '\n'` but
a quick look into the `whoami` manpage shows that it is equivalent to
calling `id -un` so we can just use `id -unz` instead and avoid having
to start a second process. This is definitely an unnecessary optimization.